@@ -2,44 +2,62 @@
   "name": "exci-maa",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "type-check": "tsc --noEmit"
   },
   "dependencies": {
+    "@clerk/clerk-react": "^5.51.0",
+    "@tanstack/react-query": "^5.0.0",
+    "axios": "^1.6.0",
+    "chart.js": "^4.4.0",
     "clsx": "^2.0.0",
     "formik": "^2.4.5",
     "framer-motion": "^10.16.16",
+    "i18next": "^24.2.3",
+    "i18next-browser-languagedetector": "^8.2.0",
     "leaflet": "^1.9.4",
     "lucide-react": "^0.294.0",
-    "multer": "^1.4.5-lts.1",
-    "next": "14.0.4",
-    "next-auth": "^4.24.5",
-    "next-intl": "^3.3.0",
-    "nodemailer": "^6.9.7",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^19.1.0",
+    "react-calendar": "^4.6.0",
+    "react-chartjs-2": "^5.2.0",
+    "react-dom": "^19.1.0",
+    "react-dropzone": "^14.2.3",
     "react-hook-form": "^7.62.0",
+    "react-hot-toast": "^2.4.1",
+    "react-i18next": "^15.7.4",
+    "react-intersection-observer": "^9.5.3",
     "react-leaflet": "^4.2.1",
+    "react-modal": "^3.16.1",
+    "react-router-dom": "^6.30.1",
+    "react-tooltip": "^5.21.4",
     "stripe": "^14.9.0",
     "tailwind-merge": "^2.2.0",
+    "typescript-eslint": "^8.46.0",
     "yup": "^1.4.0"
   },
   "devDependencies": {
-    "@types/google.maps": "^3.58.1",
     "@types/leaflet": "^1.9.8",
-    "@types/multer": "^1.4.11",
     "@types/node": "^20",
-    "@types/nodemailer": "^6.4.14",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "^19.0.0",
+    "@types/react-calendar": "^4.1.0",
+    "@types/react-dom": "^19.0.0",
+    "@types/react-modal": "^3.16.0",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@vitejs/plugin-react": "^4.0.0",
     "autoprefixer": "^10.0.1",
-    "eslint": "^8",
-    "eslint-config-next": "14.0.4",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
-    "typescript": "^5"
+    "typescript": "^5.0.2",
+    "vite": "^6.3.6"
   }
 }
